refactor(ProfitCalculator): 오버플로우 방지를 위해 revenue를 long으로 선언하라

+++ src/main/java/lotto/utility/ProfitCalculator.java
@@ -4,7 +4,7 @@ import java.util.Map;
 
 public class ProfitCalculator {
     public static double calculate(int purchaseCost, Map<String, Integer> matchedCount) {
-        int revenue = matchedCount.get("threeMatched") * LottoPrizeEnum.THREE_MATCHED_PRIZE.getAmount()
+        long revenue = matchedCount.get("threeMatched") * LottoPrizeEnum.THREE_MATCHED_PRIZE.getAmount()
                 + matchedCount.get("fourMatched") * LottoPrizeEnum.FOUR_MATCHED_PRIZE.getAmount()
                 + matchedCount.get("fiveMatched") * LottoPrizeEnum.FIVE_MATCHED_PRIZE.getAmount()
                 + matchedCount.get("fiveWithBonusMatched") * LottoPrizeEnum.FIVE_WITH_BONUS_MATCHED_PRIZE.getAmount()
